requirements.txt
PySide6>=6.5.0
numpy>=1.24.0
shapely>=2.0.0
pyvista>=0.42.0
vtk>=9.2.0
ezdxf>=1.0.0
build123d>=0.5.0
ocp-tessellate=3.10
pymeshlab
meshlib
pyransac3d
open3d
scipy
